Use a Hyrax configuration option to set the system virus scanner
Previously, we used `Hyrax.primary_work_type` to set the system wide virus
scanner. There appears to be  no reason for this to be configurable on a
per-work basis; we always use the "primary work type" (i.e. whichever one
happens to be first in the registered work types list!) to choose the virus
scanner.

Using a system-wide configuration is much more clear. The default for this
configuration is to fall back on the `primary_work_type`'s configured scanner.