refactor FileMetadata#use to #type
Instead of putting the use in a use attributes, conversion from AF File copies over all types as they exist in AF into the types attribute of FileMetadata.  To find the use, the :include? method is used to check for a specific use in the types.  This avoids data duplication by not having both type and use.  And it avoids data integrity issues when we aren’t able to correctly identify the use out of the array of types.

Co-authored-by: cjcolvar <cjcolvard@indiana.edu>