Remove `FileSet` type convienence methods
Since `#type` is located in `FileMetadata` and there's a convenient custom
query, these methods mostly function to provide `.first`. My feeling is that
this is a bit obfuscating; why not encourage queriers to handle the edge case
where there are multiple matches in context appropriate ways instead? In some
cases that might be `.first`, but not always.

This also has the benefit of removing some business logic from the model.