Refactor FileActor handling of PCDM Use vocabulary
For valkyrie normalization, take advantage for the new
`Hyrax::FileMetadata::Use.uri_for` module method to avoid reproducing
logic. When the argument is already a URI, just return it to avoid casting
unnecessarily.

In the ActiveFedora normalizer, `RDF::URI` already implements equality
appropriately, so `#casecmp` isn't needed and we can flatten to a simple switch
statement.