Ensuring CSV usage has same namespace consideration
This is a bit of a stab in the dark. However, we had a failed Ruby
build with the following message:

```console
1) Hyrax::FileSetCSVService when specifying terms and separator csv
     Failure/Error:
       ::CSV.generate do |csv|
         csv << terms.map do |term|
           values = file_set.send(term)
           values = values.respond_to?(:to_a) ? values.to_a : [values] # make sure we have an array
           values.join(multi_value_separator)
         end
       end

     NameError:
       uninitialized constant CSV
```

What I suspect to be in play is that the corresponding spec does not
similarly namespace the `CSV` constant. The production ruby file uses
`::CSV` whereas the spec ruby file uses `CSV`. This commit seeks to use
the same namespacing, hopefully reducing an intermittent error.

See the following  CircleCI build for the failures:

-  https://circleci.com/gh/samvera/hyrax/17063#tests/containers/3