Publish an `object.metadata.updated` event on transaction saves
When using the `Transactions::Step::Save` step, publish an event when the object
has saved.

The event payload wants to know which User is responsible for the metadata
update. For now, we assume this is the depositor, though that's likely not
really true in all cases.

The missing Step Arguments feature from `Dry::Transaction` may be a sticking
point here. Is it worth adding that back in our local interfaces, or are there
other patterns worth exploring?