Attempting to quash SQLite3::BusyException
On occasion, our specs encounter a `SQLite3::BusyException` exception.
This commit, based on some Google-fu (actually DuckDuckGo-ing but
that's not yet in our vernacular), I found a few references that
involve increasing the `RAILS_MAX_THREADS` environment variable and/or
increasing the database timeout.

In CircleCI, for SHA 600753e0a45f4a5682da0b8188063f8178954c11 and rspec
seed 3048, we encountered the problem. I don't assume that we can
duplicate this behavior with that configuration.

References

* https://stackoverflow.com/questions/7154664/ruby-sqlite3busyexception-database-is-locked