Tidying up RSpec/AnyInstance filename list
With any of the .rubocop_fixme.yml entries, it makes sense to revisit
on occassion. The idea being that we collectively attempt to reduce the
number of Rubocop exclusions; We will not get to 0 exclusions.

This commit reflects moving a little closer to the vision, and ensures
that the files no longer excluded operate under the same aspirational
rules.