Replacing regular expression with string comparison
Based on benchmarking, string comparison is faster than regular
express. I also believe that `include?(",")` reads better than
mentally parsing a regular expression; It is certainly more "new
developer" friendly.

Here is the benchmark that I ran locally to test performance:

```ruby
require 'benchmark'
n = 30_000_000
Benchmark.bm do |x|
  x.report { n.times { ",".include?(",") } }
  x.report { n.times { "," =~ /,/ } }
end
```

And here is the result:

```log
    user     system      total        real
4.250284   0.011629   4.261913 (  4.277509)
7.416814   0.021451   7.438265 (  7.467298)
```

And as a bonus, when attempting to bump Rubocop's `TargetRubyVersion`
from `2.3` to `2.4`, the one violation was that we were using `=~`.
Rubocop recommended that we switch to `match?` for regular
expressions (instead of `=~`).

And as an additional bonus, we were already using `include?(',')`
throughout the changed file. So this commit aligns the implementation
details of each of those points in the code.