Favoring `match?` over `=~`
Later versions of Rubocop, namely `TargetRubyVersion: 2.4` begin
flagging `=~` as something to replace. I suspect the warning is related
to `=~` pushing matching capture (e.g. the portion between `(` and `)`)
into the `$1` (and $2 if there are 2 captures). The `$1` and so on are
not thread safe.

As a matter of practice, it makes sense to move away from `=~` and into
the prefered `match?` method usage.