Introduce optional step arguments for Transactions
Adds a `Dry::Transaction`-like step argument setup via
`with_step_arguments`. When no arguments are passed for a given step, we pass
only the value wrapped by the previous step. But when step arguments are given,
we pass them into the specified step as supplemental data.

This makes it easy for individual steps to require peices of information without
other steps having to handle them. In the actor stack, we struggled with
environment parameters mutating throughout the stack, causing brittle order and
behavior dependendencies throughout the stack. Step arguments should avoid this
pitfall by ensuring callers pass step-specific data only to the step that
requires them.

To aid debugging and prevent running misconfigured transactions, we fail eagerly
when arguments are passed for steps that don't exist in the transaction.