Add resource to collection with create Transaction
Introduces a new transaction step to add an object to collections. This
step only uses `:collection_ids`, never actually requiring the collection to be
retrieved.

Issues around validity of a change that adds a work to a collection that doesn't
exist and about user permissions/abilities are dealt with in the corresponding
actor. We deliberately avoid them here since they seem best situated in
different parts of the codebase.

E.g. maybe the validation step should handle both of these concerns?