Set depositor in `WorkCreate` transaction
Use the new `Hyrax::Transactions::Transaction#with_step_args` interface to pass
a user to a step that sets the depositor via `ChangeSet`.

If no user is passed, this step is a no-op, and an automatic success. We trust
the `ChangeSet` to validate the presence of `depositor` if it is
required.

If a user is given, we set the `change_set.depositor` to the `::User#user_key`,
overriding any existing value.