Fix for HYRAX-#104
Align uploading of new versions with uploading of files

* Changed version upload to use jquery fileupload.
* Enforced max file size
* Added internationalization for messages

Expected behavior:

Select a file to be the new version. Selected file will upload.
File will be saved as new version and be properly ingested.
If the file is too large, it will not be uploaded and a message
in the current language will be displayed telling the user
that the file is too large.

In order to test:

Happy path:
* Go to the users dashboard.
* Select a work with a file set.
* Edit a file set.
* Go to versions tab.
* Upload a new file.
* Save the new version.

File too large:
* As above, but select a file that is larger than 500 MB.
* Message should display that the file is too large.
* When the interationalization files have been updated
* Change the language
* Repeat the above and file is too large message should be
  in the new language

Additionally:
* A file upload can be canceled (if it is slow enough)
* An uploaded file should be able to be deleted
* A second (or third, etc) file should be able to be uploaded
  and replace the prior new version