Backporting fix of logic around deleting nested membership
> When removing a collection after adding it to a work,
> CollectionsMembershipActor still adds the collection even though the
> _destroy flag is set to true.
>
> This is happening in both new/update works, in version 2.4.1, 2.7.0,
> and most likely 3.x.
>
> Steps to reproduce the past behavior
>
> 1.    Edit an existing work
> 2.    Add a collection to the work
> 3.    Remove the collection (before saving)
> 4.    Save the work

Based on this report, someone via the UI goes to a work, then chooses
to add a collection to the work. They then change their mind and mark
"delete" for that collection, then go to save. Prior to the logic that
is put in place, this would add the collection to the work.

The fix now prioritizes checking did someone say delete? If so, go
down that logic path else go down the add logic path.

Closes #4193