Renaming variable for collections_membership_actor
Renaming `little_boolean` to `boolean_type_caster`. As the originator
of the `little_boolean` variable name, I was aiming for playful. Yet
the name was a bit opaque. I believe the spirit of playfulness and
improved clarity of variable name shines through in
`boolean_type_caster`.

Why did I not squash this into the previous commit ("Renaming variable
for attach_members_actor")? Because I am intending to cherry-pick one
of the commits to apply to master. The other commit, is not a problem
in master because that file has been refactored.