add a default health check endpoint at `/healthz`
creates a health check endpoint at `/healthz` if the application has
the `okcomputer` gem installed. this gets us:

    - an application liveness check at `/healthz`
    - a relatively robust readiness check at `/healthz/all`

we choose `OkComputer` for the framework for this endpoint due to the ease of
extension, and due to an existing install base in parts of the community. we
should be able to source a reasonable set of default health checks from existing
code with relatively minor effort.

we'll want to extend this to check Fedora and Solr connections, and maybe
certain `valkyrie` features. for now, we just check the `ActiveRecord` DB and
the configured rails cache. if `MEMCACHED_HOST` is set, we check that; otherwise
we just use the `OkComputer` default cache check. some applications may need to
do special configuration to make sure the cache check is functioning properly.

introducing this now as an incomplete service, and with an extra step required
to turn on will give us a chance to iron out details and avoid stepping on the
toes of any applications with their own `healthz` endpoints.

the check can be installed during app generation by passing `false` to the
`skip-health-check` option.