refactor `Hyrax::SolrService` specs
avoid hard overwrites of Hyrax configuration in the SolrService specs.

the contract here guarantees that the configuration given in the initalizer will
be used. take advantage of that to better isolate and deduplicate the tests.

if we need to check the default setting (i don't think we do), we can always
check that the setting is correct after a generic `described_class.new`
separately from these more complex behavioral tests.