better tests and documentation for `Hyrax::SimpleSchemaLoader`
this private class is heavily used by the lightweight schema definitions for
core and basic metadata. a bit more testing and documentation seemed called for
as i was looking at this, so i put it together.

i also refactored an internal method to maybe be a bit more clear, and avoid an
`&.` operator.