remove the indexer initializer generator
the code this generator injects isn't really necessary (if we want to register
this indexer for the generic `Hyrax::Work`, we can do it in Hyrax
directly). including it complicates the upgrade process for existing
apps. instead, while we need to manually register indexers for specific
works (maybe we can do this with configuration soon, instead?), inject them into
the hyrax initializer when the work generator runs.