start to bridge the gap between HydraEditor forms and ChangeSets
for Valkyrie models, we'd like to use Reform-based `ChangeSet` instances as the
basis for forms. this is a lot simpler for "dirty tracking" than the large-scale
delegation approach used by HydraEditor. it's also more powerful, allowing these
form objects to be used to construct forms and manage attribute processing. this
latter responsibility is currently a bit scattered between controllers, service
classes, and the actor stack. there's reason to believe this shift will help us
clean that up.

however, we *also* want to avoid major controller/view side changes, so we want
these forms to support `HydraEditor::Form`-like interfaces, as well as the
specific behavior existing `Hyrax::Forms` classes.

this starts to extend the `Valkyrie::ChangeSet`/`Hyrax::ChangeSet` behavior to
support expected behavior of `HydraEditor::Form`. the expectation is that
application-side forms for valkyrie models will be defined like:

    class MonographForm < Hyrax::Forms::ResourceForm(Monograph)
      self.required_fields = [:title, :creator, :rights_statement]

      # other WorkForm-like configuration here
    end

this won't necessarily result in an attempt to *completely* reproduce the
WorkForm interface; rather, the goal is to narrow the scope of application-side
porting work required to shift to the above. especially, we want external
behavior to be as close as possible.