refactor `WorksControllerBehavior.iiif_manifest_builder`
use a more OO style for the dependency injection on `iiif_manifest_builder`.

makes the caching behavior a subclass, instead of a branch, and places the
configuration in the controller. this should make the manifest builders more
reusable for other controllers, which might decide on caching behavior in a
different way.

renames `#as_json` (which used to return a hash!) to `#manifest_for` and usees
the instance method in callers.