use helpers to i18nize configured permission levels/options
these configured options (why do they need to be configurable? how would one
configure them successfully?) need to be i18nized, but the current fix only
caches the i18n strings at app startup (or maybe when the configurations are
first hit?).

to translate on-the-fly we need to hit the i18n key while the user is in the
desired locale.

this is a stop-gap fix. probably we want to make the configuration options just
straight arrays, or remove them completely. discussion on longer term approachs
is scheduled in https://wiki.lyrasis.org/display/samvera/Samvera+Tech+Call+2020-06-05.