Addressing case where children are added again
Prior to this commit, there was no guard in place to prevent adding
children again.

The logic would check "is something marked for delete" if not, then
add it. That needed to change, because when a user would submit a
form changing keywords, then we would re-add all of the already added
members.

Related to #4301

This ports forward changes made in b201f7310f805b7e9ecd48ec0bb6838cfa7f49cd