expand the `EmbargoManager` interface to handle releasing embargos
for Valkyrie models, we provide `Hyrax::EmbargoManager` to make the embargo
lifecycle easy to work with. to date, this has only supported applying
embargoes. this adds the ability to release an embargo after its period has expired.

a concept of 'enforcement' is added, represented by `#enforced?`. this addresses
the issue of embargoes which are not current (their period is past) but have not
yet been released.

the terminology used by the `EmbargoManager` is documented.

see also: https://github.com/samvera/hydra-head/issues/511.