document and test `Hyrax::LeaseManager`
`LeaseManager` is substantially similar to `EmbargoManager`, and got in without
tests, with the idea that a refactor might quickly take place. it didn't and at
this point it needs a test harness if that's ever going to happen.

in the meanwhile. there are features we want to add to cover the remainder of
the lifecycle.