add handling for embargo/lease and agreements to the ChangeSet form
since the new valkyrie models don't have pass-through attributes for
embargo/lease settings, this form uses Reform's virtual properties to accept
this input from the existing views. when saving the form, it will be necessary
to interpret the virtual property states to determine appropriate changes to
embargoes and leases. the values are prepopulated appropriately when
`#prepopulate!` is called.

for `#agreeement_accepted` we use a similar approach, setting to `false` by
default, and to `true` when prepopulating with an existing model object.