test form views with ChangeSet style form objects--create
recent work has extended the ChangeSet style `Hyrax::ResourceForm` to support
most of the view behavior for work create forms. this adds tests of the relevant
form partials with the new form object.

the partial that handles collection membership is stubbed, since its behavior is
more complex and reproducing it in the new form should be considered separately.

likewise, tests for these partials as an edit form (for existing objects) are
left out; mainly because versions/optimistic locking becomes an issue in this
case. there's more groundwork to lay there.

to support this, a registration step is added to the new model generator. this
sets up the routes when we bulid the test app.