allow FilterByType to handle models without `#to_rdf_representation`
`FilterByType` used to rely on `.to_rdf_representation`, which is `ActiveFedora`
specific. the default implementation is just `.name` ("RDF representation" means
here "unique name referring to the model type", Class.name fits this bill fine).

Add support for Valkyrie works in this context by using `.name` as a
fallback. possibly we'll want a different name for *actual* RDF representations,
but since this is really just for the solr index, probably we can just use
`.name` forever for the new models.