support HydraEditor::Form objects in EmbargoHelper#embargo_enforced?
extends `EmbargoHelper#embargo_enforced?` to handle the case when a form object
is explictly checked for an enforced embargo, instead of its model.

rather than counting on the presence of a delegated method, we call recursively
with `resource.model`. since `#model` is core to the `HydraEditor::Form`
interface, this should be very resiliant.