add a job for prewarming a manifest cache
this job hits the manifest cache to prewarm; at least for now it's a (fast) no-op if
there is already something in the cache for the key.

reconstruct the cache keys to use `etag`. this isn't perfect, but it correlates
more closely to object changes than the solr `_version_` field, and avoids worry
about cases where a `SolrDocument#[]` lookup on `_version_` is `nil` (which are
common). for AF models, `#etag` is populated as long as the model isn't a
`#new_record?`.

we also add a prefix for the cache key, this is meant to be specific to the
caching implementation. if we change the implementation in the future in a way
that would benefit from ignoring old cache entries, we can simply change or
increment the prefix. this also serves as a namespace, which should avoid cache
key collisions.