implement primary/secondary terms in ChangeSet forms
the old Hyrax::Forms::WorkForm implements `#primary_terms` and
`#secondary_terms` (along with `#required?`) by asking inheriting classes to
set an `Array` determining each.

this was a bit error prone in practice, since it's easy to accidentally set the
same term to both settings, or to drop one altogether. this was especially an
issue because the form implemented all of basic metadata in the base
class. since the new forms have property definitions via `Twin`/`Disposable` and
these definitions can accept arbitrary data, we have the option to leverage this
to define term behavior in an extensible way.

this work uses that existing infrastructure to reimplement `#primary_terms` and
`#secondary_terms` (we've already used this infrastructure for `#required?`).

we introduce `display: true|false`, and apply it to all terms defined by a
schema configuration. we also introduce `primary: true|false` and flag existing
"primary terms". "secondary terms" become those that are not-primary, but also
flagged for display.