use a `collection_member_sort_fields` helper to manage sort options
collection show views were using the default blacklight config for
`active_sort_fields`. there's not much point in sorting by relevance, and other
changes to how these sort options present may be desirable on a per-application
basis.

to make it easy to make this kind of change, add a
`collection_member_sort_fields` helper. for now, its implementation is just a
call to the existing helper, so actual behavior is fixed for this commit.

connected to #4308.