use a specialized `IiifManifestPresenter` for manifest generation
this complete rewrite of the manifest generation process reduces the footprint
of database (Fedora/Solr) calls involved in the manifest generation process,
fixes a bug in sequence/canvas generation for manifests for nested works, and
generally separates the issue of manifest generation from the
`WorkShowPresenter` (existing code is left in place, to be deprecated and
removed as follow-up).

[`iiif_manifest`](https://www.rubydoc.info/gems/iiif_manifest/1.0.1) expects a
"presenter", with `#manifest_url`, `#manifest_metadata` `#description`,
`#sequence_rendering`, and `#work_presenters`/`#file_set_presenters`. since we
know the presenter provided will be used for the duration of a single request,
we can make heavy use of instance variables to cache children and avoid making
multiple requests for any particular data.

the rewrite also does some work to help clarify the two peices of information
the presenter needs in the request context: the `Ability` and the hostname to
use for URIs within the manifest. we provide naive defaults for both of these,
and implement the behavior more or less without regard for their
context---i.e. the whole presenter infrastructure works without them. when
calling from a Controller, we pass them in explicitly to augment the behavior
for the context. this is important because we may (do) want to reuse these
presenters in other contexts (a Job for cache pre-warming) where this data isn't
available.

the presenters themselves use `Draper::Decorator`'s `delegate_all`, so they're
basically glorified wrappers for a model. they don't care whether the model is a
SolrDocument, or powered by a primary database. a small amount of work is done
to align Valkyrie, ActiveFedora, and SolrDocument-style models, but mostly
things were already interoperable.

a refactor in `DisplaysImage` rearranges some instance-level caching there, for
easier reuse in the new presenter infrastructure.

the short story is: between this and a previous optimization in
`DisplaysImage` (#4344), manifest generation in a makeshift controlled
benchmarking environment sped up 10-fold for works of modest size. it seems
roughly linear between 10 and 750 child FileSets. i'm not confident enough in
that environment (especially its behavior WRT caching) to call this "data" or to
provide anything more solid than this statement, but there's good reason to
believe this will be quite a bit faster/more scalable than the status-quo.