wings: round trip Valkyrie-native models, even when unregistered
this is the beginning of an effort to better generalize model conversion
handling in `Wings` to handle native models without an explicit,
application-side `ActiveFedora` equivalent.

the bar we're seeking to clear here is: get a resource/work class generated with
the Hyrax valkyrie generator to round trip in Wings without any specialized
AF-related handling. this doesn't yet try to assess the quality of metadata
round tripping, but that work should follow.