wings: support optimistic locking for Valkyrie resources
adds an optimistic locking implementation to `Wings` using Fedora's
`Last-Modified` property and etags.

the existing optimistic locking in the Actor stack counts on etags, so including
them in the lock tokens has immediate value. we'll need a strategy for removing
them later.

for now, this is a fairly naive implementation for early consideration.