pin reform to v2.2.x and under
the Reform v2.3.1 release broke `reform-rails` and
`valkyrie`. https://github.com/trailblazer/reform/issues/508

the plan on the `trailblazer` side appears to be to roll forward by releasing a
new, compatible version of `reform-rails`. to fix the build in the meanwhile,
pin `reform` to a working release.

if `reform-rails` is released from a current release candidate, that should also
patch this issue; see https://github.com/samvera/hyrax/pull/4332. assuming that
happens before Hyrax 3.0.0, we can bump both the `reform` and `reform-rails`
depend on the latest versions directly.