wings: remove custom converter handling for apply_depositor_metadata
both the depositor attribute and the ACLs are now handled robustly by the core
converter logic. using this custom `#apply_depositor_metadata` method, which
edits both the property and the ACLs, runs the risk of mucking them up. there's
no need for it here anymore, so let's get it out of there.