allow apps to forego a custom ModelActor
in the past we fail at runtime with NameError if an application fails to provide
a custom actor for its local models. this isn't particularly forgiving or
informative, and we probably don't want to require actors forever (see
the dry-monad transaction infrastructure). besides many applications don't
provide work-specific actor customizations, so requiring them to provide an
actor per-work might be silly.

instead we provide a new `NullActor`, which just does nothing and passes along
to the next actor. if a model actor is missing, log the miss and use `NullActor`
instead.